write API for inserting data into tables

+++ V-Tracker/www/scripts/storage.js
@@ -14,77 +14,46 @@ var storage = {
 	init:function() {
 		db = window.openDatabase("VTracker_db", "1.00", "V-Tracker DB", 2 * 1024*1024); //open a 2MBs database
 
-		storage.createTable('consolelog','CONSOLELOG');
+		storage.createTable(consoleLogObj,'CONSOLELOG');
 		
 		storage.ready = true;
-		consoleLog.add("Database opened. Storage initiated. Storage ready: " + storage.ready);
+		console.log("Database opened. Storage initiated. Storage ready: " + storage.ready);
 	},
 	//******************************** END INITIALISE STORAGE *********************************//
 	
 	//************************************ RESET DATABASES ************************************//
 	reset:function() {
 		storage.ready = false;
-		storage.dropTable(storage.dbTables)	
-		
+		storage.dropTable(storage.dbTables); //drop all tables we know about at this stage
 		console.log("Database tables reset! Storage ready: " + storage.ready);
-		
-		storage.init();
 	},
 	//********************************** END RESET DATABASES **********************************//
 	
 	//************************************* CREATE TABLE **************************************//
-	createTable:function(type, tableName) {
-		var dbFields = "id unique";
-		
-		switch(type) {
-			case "geolocation":
-				for (var i in geolocationObj.data) {
-				  	dbFields = dbFields + "," + i;
-				}
-				break;
-				
-			case "compass":
-				for (var i in compassObj.data) {
-				  	dbFields = dbFields + "," + i;
-				}
-				break;	
-			
-			case "accelerometer":
-				for (var i in accelerometer.data) {
-				  	dbFields = dbFields + "," + i;
-				}
-				break;
-				
-			case "gyroscope":
-				for (var i in gyroscopeObj.data) {
-				  	dbFields = dbFields + "," + i;
-				}
-				break;
-				
-			case "consolelog":
-				dbFields = "id, Timestamp, Message";
-				break;
-			
-			default:
-				console.log("db table: \'" + tableName + "\' could not be created");
-				return
+	createTable:function(parentObject, tableName) {
+		//Specify schema for table based on the object's 'data' properties
+		var dbFields = "id integer primary key autoincrement";
+		for (var i in parentObject["data"]) {
+			dbFields = dbFields + "," + i;
 		}
 		
+		//create table
 		db.transaction(function (tx) {
 					   tx.executeSql('CREATE TABLE IF NOT EXISTS ' + tableName + ' (' + dbFields + ');');
 					   }, storage.errorCB, storage.successCB);
 		
-		storage.getDBTables();				   		   
+		storage.getDBTables();   		   
 	},
 	
 	//*********************************** END CREATE TABLE ************************************//
 	
 	//************************************** DROP TABLE ***************************************//
 	dropTable:function(tableNames) {
-		db.transaction(function (tx) {for (var i=0;i<tableNames.length;i++) {
+		db.transaction(function (tx) { for (var i=0;i<tableNames.length;i++) {
+											if(tableNames[i] == "sqlite_sequence") {continue;}
 											tx.executeSql('DROP TABLE IF EXISTS ' + tableNames[i] + ';');
 											}
-									  }, [], storage.errorCB, storage.init);
+									  }, storage.errorCB, storage.init);
 		storage.getDBTables();
 	},
 	//************************************ END DROP TABLE *************************************//
@@ -97,104 +66,87 @@ var storage = {
 	successCB:function(toStore) {
 		//=== do not call the consoleLog object in here, because you'll get an infinite loop on success ===//
 	    
-		//console.log("SQL processed successfully!");
+	    //console.log("SQL processed successfully!");
 	},
 	//********************************** END ERROR HANDLING ***********************************//
 	
 	//******************************** INSERT DATA INTO TABLES ********************************//
-	rowIdCounter: {
-		consoleLogTable: 0,
-		accelerometerTable: 0,
-		compassTable: 0,
-		gyroscopeTable: 0,
-		geolocationTable: 0
-	},
-	
-	updateTable:function(tableName, data) {
+	insertInTable:function(tableName, dataType, data) {
+		switch(dataType) {
+			//console data
+			case "consoleData":
+				var message = data;
+				var toStore = '"' + new Date(new Date().getTime()) + '",' +
+				              '"' + message + '"';
+				
+				db.transaction(function (tx) {tx.executeSql('INSERT INTO CONSOLELOG VALUES (null,' + toStore + ');');
+											 }, function (error) { //onError
+										   			console.log("Error logging: " + message + " | Details: " + error.code + ", " + error.message);
+							   				  					}, storage.successCB);
+				return;
+			
+			//accelerometer data
+			case "accelerationData":
+				var accelerationData = data;
+				var toStore =	'"' + new Date(accelerationData.timestamp) + '",' +
+								'"' + accelerationData.x          + '",' +
+								'"' + accelerationData.y          + '",' +
+								'"' + accelerationData.z          + '"';
+				break;
+
+			//compass data
+			case "compassData":
+				var compassHeading = data;
+				var toStore = 	'"' + new Date(compassHeading.timestamp)  + '",' +               
+								'"' + compassHeading.magneticHeading      + '",' + 
+								'"' + compassHeading.trueHeading    	  + '",' + 
+								'"' + compassHeading.headingAccuracy      + '"';
+				break;
+
+			//gyroscope data
+			case "gyroData":
+				var gyroData = data;
+				var toStore = 	'"' + new Date(gyroData.timestamp) 	+ '",' +               
+								'"' + gyroData.alpha    	+ '",' +
+								'"' + gyroData.beta     	+ '",' +
+								'"' + gyroData.gamma    	+ '"';
+				break;
+							
+			//geolocation data
+			case "geolocationData":
+				var position = data;
+				var toStore =	'"' + new Date(position.timestamp)      + '",' +
+								'"' + position.coords.latitude          + '",' +
+								'"' + position.coords.longitude         + '",' +
+								'"' + position.coords.altitude          + '",' +
+								'"' + position.coords.accuracy          + '",' +
+								'"' + position.coords.altitudeAccuracy  + '",' +
+								'"' + position.coords.heading           + '",' +
+								'"' + position.coords.speed             + '"';
+				break;
+			
+			default: console.log("Error: Could not format data for storage!");
+		}
+							
 		db.transaction(function (tx){
 	  					 tx.executeSql('SELECT name, sql FROM sqlite_master WHERE type="table" AND name = "' + tableName + '";', [], function (tx, results) {
 							  var columnParts = results.rows.item(0).sql.replace(/^[^\(]+\(([^\)]+)\)/g, '$1').split(',');
-							  console.log(columnParts);
-									 }, storage.errorCB);
-					   }, storage.errorCB, storage.successCB);
-	},	
-	
-	//can't use this. in here...
-	updateSQLTable: {
-		//console log table
-		consoleLog:function(message) {
-			var toStore = '"' + new Date(new Date().getTime()) + '",' + 
-			              '"' + message + '"';
-			
-			db.transaction(function (tx) {
-										tx.executeSql('INSERT INTO CONSOLELOG VALUES (' + storage.rowIdCounter.consoleLogTable + ',' + toStore + ');');
-										storage.rowIdCounter.consoleLogTable++;
-										}, function (error) { //onError
-									   			console.log("Error logging: " + message + " | Details: " + error.code + ", " + error.message);
-						   				  					}, storage.successCB(toStore));
-		},
-		
-		//accelerometer table
-		accelerometer:function(accelerationData) {
-			var toStore =	'"' + new Date(accelerationData.timestamp) + '",' +
-							'"' + accelerationData.x          + '",' +
-							'"' + accelerationData.y          + '",' +
-							'"' + accelerationData.z          + '"';
-			
-			db.transaction(function (tx) {
-										tx.executeSql('INSERT INTO ACCELEROMETER VALUES (' + storage.rowIdCounter.accelerometerTable + ',' + toStore + ');');
-										storage.rowIdCounter.accelerometerTable++;
-										}, storage.errorCB, storage.successCB(toStore));
-		},
-		
-		//compass table
-		compass:function(compassHeading) {
-			var toStore = 	'"' + new Date(compassHeading.timestamp)  + '",' +               
-							'"' + compassHeading.magneticHeading      + '",' + 
-							'"' + compassHeading.trueHeading    	  + '",' + 
-							'"' + compassHeading.headingAccuracy      + '"';
-			
-			db.transaction(function (tx) {
-										tx.executeSql('INSERT INTO COMPASS VALUES (' + storage.rowIdCounter.compassTable + ',' + toStore + ');');
-										storage.rowIdCounter.compassTable++;
-										}, storage.errorCB, storage.successCB(toStore));
-		},
-		
-		//gyroscope table
-		gyroscope:function(gyroData) {
-			var toStore = 	'"' + new Date(gyroData.timestamp) 	+ '",' +               
-							'"' + gyroData.alpha    	+ '",' +
-							'"' + gyroData.beta     	+ '",' +
-							'"' + gyroData.gamma    	+ '"';
-							
-			
-			db.transaction(function (tx) {
-										tx.executeSql('INSERT INTO GYROSCOPE VALUES (' + storage.rowIdCounter.gyroscopeTable + ',' + toStore + ');');
-										storage.rowIdCounter.gyroscopeTable++;
-										}, storage.errorCB, storage.successCB(toStore));
-		},
-		
-		//geolocation table
-		geolocation:function(position) {	
-			var toStore =	'"' + new Date(position.timestamp)      + '",' +
-							'"' + position.coords.latitude          + '",' +
-							'"' + position.coords.longitude         + '",' +
-							'"' + position.coords.altitude          + '",' +
-							'"' + position.coords.accuracy          + '",' +
-							'"' + position.coords.altitudeAccuracy  + '",' +
-							'"' + position.coords.heading           + '",' +
-							'"' + position.coords.speed             + '"';
-		
-			db.transaction(function (tx) {
-										tx.executeSql('INSERT INTO GEOLOCATION VALUES (' + storage.rowIdCounter.geolocationTable + ',' + toStore + ');');
-										storage.rowIdCounter.geolocationTable++;
-										}, storage.errorCB, storage.successCB(toStore));
-		}
+							  //console.log(columnParts);
+							  
+							  //----- INSERT data into table, if we have the correct number of measurements
+							  if (data.length == columnParts.length-1) {
+							  	db.transaction(function (tx) {
+										tx.executeSql('INSERT INTO ' + tableName + ' VALUES (null,' + toStore + ');');
+								}, storage.errorCB, storage.successCB);
+							  }
+							  //----- END
+							  
+						}, storage.errorCB);
+		}, storage.errorCB, storage.successCB);
 	},
 	//****************************** END INSERT DATA INTO TABLES ******************************//
 	
 	//************************ RETRIEVING TABLES AND LENGTHS FROM DB **************************//	
-	queryCounter: 0,
 	dbTables: [], //declare array of database tables
 	
 	getDBTables:function() {
@@ -210,6 +162,7 @@ var storage = {
 					   }, storage.errorCB, function() {console.log("dbTables updated: " + storage.dbTables)});
 	},
 	
+	queryCounter: 0,
 	getDBTableLengths:function() {
 		//if this is the first run, then set the output window
 		if (storage.queryCounter == 0) {$('#databases').append('<p>started...</p>');}
@@ -238,8 +191,8 @@ var storage = {
 		db.transaction(function (tx){
 						   tx.executeSql(tableQuery, [], function (tx, results) {
 									 	$('#databases').append('<p>' + tableName + ' length: ' + results.rows.length + '</p>');
-									 }, this.errorCB);
-						   }, this.errorCB, this.getDBTableLengths);	
+									 }, storage.errorCB);
+						   }, storage.errorCB, storage.getDBTableLengths);	
 	}
 	//********************** END RETRIEVING TABLES AND LENGTHS FROM DB ************************//		
 }
